Feature/config changes

+++ server/src/seeds/cleanDb.ts
@@ -1,19 +1,13 @@
-// import Question from '../models/index.js';
-import db from '../config/connection.js';
 import models from '../models/index.js';
+import db from '../config/connection.js';
 
-export default async (modelName: keyof typeof models , collectionName: string) => {
+export default async (modelName: "Question", collectionName: string) => {
   try {
-    let modelExists = [];
-    if (models[modelName] && models[modelName].db) {
-      if (models[modelName].db.db) {
-        modelExists = await models[modelName].db.db.listCollections({
-          name: collectionName
-        }).toArray();
-      }
-    }
+    let modelExists = await models[modelName].db.db?.listCollections({
+      name: collectionName
+    }).toArray()
 
-    if (modelExists.length) {
+    if (modelExists && modelExists.length) {
       await db.dropCollection(collectionName);
     }
   } catch (err) {
